+++ src/options.cpp
@@ -30,6 +30,9 @@ std::map<int, std::string> Option::s_header
 
 std::map<int, std::string> Option::s_measureNumber
     = { { MEASURENUMBER_system, "system" }, { MEASURENUMBER_interval, "interval" } };
+    
+std::map<int, std::string> Option::s_systemDivider
+    = { { SYSTEMDIVIDER_none, "none" }, { SYSTEMDIVIDER_left, "left" }, { SYSTEMDIVIDER_left_right, "left-right"} };
 
 //----------------------------------------------------------------------------
 // Option
@@ -734,6 +737,10 @@ Options::Options()
     m_stemWidth.SetInfo("Stem width", "The stem width");
     m_stemWidth.Init(0.20, 0.10, 0.50);
     this->Register(&m_stemWidth, "stemWidth", &m_generalLayout);
+    
+    m_systemDivider.SetInfo("System divider", "The display of system dividers");
+    m_systemDivider.Init(SYSTEMDIVIDER_left, &Option::s_systemDivider);
+    this->Register(&m_systemDivider, "systemDivider", &m_generalLayout);
 
     m_tieThickness.SetInfo("Tie thickness", "The tie thickness in MEI units");
     m_tieThickness.Init(0.5, 0.2, 1.0);